调整默认参数

+++ scripts/inference.py
@@ -8,6 +8,13 @@ import glob
 import pickle
 from tqdm import tqdm
 import copy
+import sys
+sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
+
+
+# set ffmpeg path
+# tmp_cmd = f"export FFMPEG_PATH=../ffmpeg"
+# os.system(tmp_cmd)
 
 from musetalk.utils.utils import get_file_type,get_video_fps,datagen
 from musetalk.utils.preprocessing import get_landmark_and_bbox,read_imgs,coord_placeholder
@@ -19,6 +26,8 @@ import shutil
 audio_processor, vae, unet, pe = load_all_model()
 device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
 timesteps = torch.tensor([0], device=device)
+
+
 
 @torch.no_grad()
 def main(args):
@@ -67,7 +76,7 @@ def main(args):
         #print(input_img_list)
         ############################################## extract audio feature ##############################################
         whisper_feature = audio_processor.audio2feat(audio_path)
-        whisper_chunks = audio_processor.feature2chunks(feature_array=whisper_feature,fps=fps)
+        whisper_chunks  = audio_processor.feature2chunks(feature_array=whisper_feature,fps=fps)
         ############################################## preprocess input image  ##############################################
         if os.path.exists(crop_coord_save_path) and args.use_saved_coord:
             print("using extracted coordinates")
@@ -88,35 +97,36 @@ def main(args):
             x1, y1, x2, y2 = bbox
             crop_frame = frame[y1:y2, x1:x2]
             crop_frame = cv2.resize(crop_frame,(256,256),interpolation = cv2.INTER_LANCZOS4)
-            latents = vae.get_latents_for_unet(crop_frame)
+            latents    = vae.get_latents_for_unet(crop_frame)
             input_latent_list.append(latents)
     
         # to smooth the first and the last frame
         frame_list_cycle = frame_list + frame_list[::-1]
         coord_list_cycle = coord_list + coord_list[::-1]
         input_latent_list_cycle = input_latent_list + input_latent_list[::-1]
+        
+        
         ############################################## inference batch by batch ##############################################
         print("start inference")
-        video_num = len(whisper_chunks)
+        video_num  = len(whisper_chunks)
         batch_size = args.batch_size
-        gen = datagen(whisper_chunks,input_latent_list_cycle,batch_size)
+        gen        = datagen(whisper_chunks,input_latent_list_cycle,batch_size)
         res_frame_list = []
         for i, (whisper_batch,latent_batch) in enumerate(tqdm(gen,total=int(np.ceil(float(video_num)/batch_size)))):
             audio_feature_batch = torch.from_numpy(whisper_batch)
-            audio_feature_batch = audio_feature_batch.to(device=unet.device,
-                                                         dtype=unet.model.dtype) # torch, B, 5*N,384
+            audio_feature_batch = audio_feature_batch.to(device=unet.device, dtype=unet.model.dtype) # torch, B, 5*N,384
             audio_feature_batch = pe(audio_feature_batch)
-            latent_batch = latent_batch.to(dtype=unet.model.dtype)
+            latent_batch        = latent_batch.to(dtype=unet.model.dtype)
             
             pred_latents = unet.model(latent_batch, timesteps, encoder_hidden_states=audio_feature_batch).sample
-            recon = vae.decode_latents(pred_latents)
+            recon        = vae.decode_latents(pred_latents)
             for res_frame in recon:
                 res_frame_list.append(res_frame)
                 
         ############################################## pad to full image ##############################################
         print("pad talking image to original video")
         for i, res_frame in enumerate(tqdm(res_frame_list)):
-            bbox = coord_list_cycle[i%(len(coord_list_cycle))]
+            bbox      = coord_list_cycle[i%(len(coord_list_cycle))]
             ori_frame = copy.deepcopy(frame_list_cycle[i%(len(frame_list_cycle))])
             x1, y1, x2, y2 = bbox
             try:
@@ -125,7 +135,7 @@ def main(args):
 #                 print(bbox)
                 continue
             
-            combine_frame = get_image(ori_frame,res_frame,bbox)
+            combine_frame = get_image(ori_frame, res_frame, bbox)
             cv2.imwrite(f"{result_img_save_path}/{str(i).zfill(8)}.png",combine_frame)
 
         cmd_img2video = f"ffmpeg -y -v warning -r {fps} -f image2 -i {result_img_save_path}/%08d.png -vcodec libx264 -vf format=rgb24,scale=out_color_matrix=bt709,format=yuv420p -crf 18 temp.mp4"
@@ -136,26 +146,22 @@ def main(args):
         print(cmd_combine_audio)
         os.system(cmd_combine_audio)
         
-        os.remove("temp.mp4")
+        if os.path.exists("temp.mp4"):
+            os.remove("temp.mp4")
         shutil.rmtree(result_img_save_path)
         print(f"result is save to {output_vid_name}")
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
-    parser.add_argument("--inference_config", type=str, default="configs/inference/test_img.yaml")
+    parser.add_argument("--inference_config", type=str, default="/home/zhangss/work/MuseTalk/configs/inference/test.yaml")
     parser.add_argument("--bbox_shift", type=int, default=0)
     parser.add_argument("--result_dir", default='./results', help="path to output")
 
-    parser.add_argument("--fps", type=int, default=25)
-    parser.add_argument("--batch_size", type=int, default=8)
+    parser.add_argument("--fps",             type=int, default=25)
+    parser.add_argument("--batch_size",      type=int, default=8)
     parser.add_argument("--output_vid_name", type=str, default=None)
-    parser.add_argument("--use_saved_coord",
-                        action="store_true",
-                        help='use saved coordinate to save time')
-    parser.add_argument("--use_float16",
-                        action="store_true",
-                        help="Whether use float16 to speed up inference",
-    )
+    parser.add_argument("--use_saved_coord", default=True, help='use saved coordinate to save time')
+    parser.add_argument("--use_float16",     default=True, help="Whether use float16 to speed up inference",)
 
     args = parser.parse_args()
     main(args)
